// src/app/main/admin/events/model/event.model.ts
export class EventDetail
{
    id?: string;
    name?: string;
    isactive?: string;
    isspeed?: string;
    speedcondition?: string;
    speed?: string; 
    isoffhours?: string; 
    offhourstart?: string; 
    offhourend?: string; 
    islatestart?: string;
    startofday?: string; 
    islongstop?: string; 
    stoplength?: string; 
    isnodata?: string; 
    nodataintervalon?: string; 
    nodataintervaloff?: string; 
    isidle?: string; 
    idletime?: string; 
    iszone?: string; 
    zoneid?: string; 
    zone?: string; 
    zonecondition?: string; 
    ispoi?: string; 
    poiid?: string ;
    poi?: string ;
    poicondition?: string ;
    ishardware?: string; 
    devicecode?: string; 
    isroute?: string; 
    routedeviation?: string; 
    isfuel?: string; 
    fuelvalue?: string; 
    fuelunits?: string; 
    fuelcondition?: string; 
    fueltank?: string; 
    istemp?: string; 
    tempmin?: string; 
    tempmax?: string; 
    tempunits?: string; 
    tempchange?: string; 
    tempsensor?: string; 
    tempcondition?: string; 
    isdigitalin?: string; 
    inputon?: string; 
    inputoff?: string; 
    companyid?: string; 
    groupid?: string; 
    isfullcompany?: string; 
    isfullgroup?: string; 
    istow?: string;
    method?: string;
    conncode?: string;
    userid?: string;
    company?: string;
    group?: string; 

    constructor(event_detail?)
    {
       event_detail = event_detail || {}
       this.conncode = event_detail.conncode || '';
       this.userid = event_detail.userid || '0';
       this.method = event_detail.method || '';
       this.id = event_detail.id || '0';
       this.name = event_detail.name || '';
       this.isactive = event_detail.isactive || 'false';
       this.isspeed = event_detail.isspeed || 'false';
       this.speedcondition = event_detail.speedcondition || '0';
       this.speed = event_detail.speed || '0';
       this.isoffhours = event_detail.isoffhours || 'false';
       this.offhourstart = event_detail.offhourstart || '0';
       this.offhourend = event_detail.offhourend || '0';
       this.islatestart = event_detail.islatestart || 'false';
       this.startofday = event_detail.startofday || '0';
       this.islongstop = event_detail.islongstop || 'false';
       this.stoplength = event_detail.stoplength || '0';
       this.isnodata = event_detail.isnodata || '0';
       this.nodataintervalon = event_detail.nodataintervalon || '0';
       this.nodataintervaloff = event_detail.nodataintervaloff || '0';
       this.isidle = event_detail.isidle || 'false';
       this.idletime = event_detail.idletime || '0';
       this.iszone = event_detail.iszone || '0';
       this.zoneid = event_detail.zoneid || '0';
       this.zone = event_detail.zone || '';
       this.zonecondition = event_detail.zonecondition || '0';
       this.ispoi = event_detail.ispoi || 'false';
       this.poiid = event_detail.poiid || '0';
       this.poi = event_detail.poi || '';
       this.poicondition = event_detail.poicondition || '0';
       this.ishardware = event_detail.ishardware || 'false';
       this.devicecode = event_detail.devicecode || '';
       this.isroute = event_detail.isroute || 'false';
       this.routedeviation = event_detail.routedeviation || '0';
       this.isfuel = event_detail.isfuel || '0';
       this.fuelvalue = event_detail.fuelvalue || '0';
       this.fuelunits = event_detail.fuelunits || '0';
       this.fuelcondition = event_detail.fuelcondition || '0';
       this.fueltank = event_detail.fueltank || '0';
       this.istemp = event_detail.istemp || 'false';
       this.tempmin = event_detail.tempmin || '0';
       this.tempmax = event_detail.tempmax || '0';
       this.tempunits = event_detail.tempunits || '0';
       this.tempchange = event_detail.tempchange || '0';
       this.tempsensor = event_detail.tempsensor || '0';
       this.tempcondition = event_detail.tempcondition || '0';
       this.isdigitalin = event_detail.isdigitalin || 'false';
       this.inputon = event_detail.inputon || '0';
       this.inputoff = event_detail.inputoff || '0';
       this.companyid = event_detail.companyid || '0';
       this.groupid = event_detail.groupid || '0';
       this.isfullcompany = event_detail.isfullcompany || 'false';
       this.isfullgroup = event_detail.isfullgroup || 'false';
       this.istow = event_detail.istow || 'false';
    }
}